Mise à jour
- Ajout du calcul de la boundingbox pour les mesh dans pb_steploadobjfile
- Suppression d'une classe plus utilisé
- Ajout de constructeur de recopie pour les itérateurs
- Si lors de l'ajout/modification d'un point on passe un système de coordonnée à NULL, il prend le système de coordonnée par défaut qui ne fait pas de translation

git-svn-id: http://rdinnovation.onf.fr/svn/computree@295 0d0f61bb-bbfd-4940-b5ed-fa75979e1372

+++ pluginshared/ct_pointcloud/tools/abstract/ct_abstractundefinedsizepointcloud.h
@@ -43,7 +43,7 @@ public:
      * @brief Add a point to this cloud. You must pass the index of the coordinate system that will be used to transform the point in
      *        float values.
      */
-    virtual void addPoint(const CT_Point &point, const size_t &coordinateSystemGlobalIndex) = 0;
+    virtual void addPoint(const CT_Point &point, const GLuint &coordinateSystemGlobalIndex) = 0;
 
     /**
      * @brief Add a point to this cloud. You must pass a coordinate system that will be set for this point
@@ -53,7 +53,7 @@ public:
     /**
      * @brief Add a point to this cloud. You must pass the index of the coordinate system that will be set for this point
      */
-    virtual void addInternalPoint(const CT_PointData &point, const size_t &coordinateSystemGlobalIndex) = 0;
+    virtual void addInternalPoint(const CT_PointData &point, const GLuint &coordinateSystemGlobalIndex) = 0;
 
 protected:
     /**
